Fixed saver relative paths for `latest_checkpoint`
This would be cleaner if we made all paths listed in the "latest"
file relative to the its directory, allowing the removal of the
added `os.path.isabs` checks.

That would make the `os.join` in `saver.latest_checkpoint` much less
surprising.

But at least this way, there is no effect on currently working code.

Fixes #571
Change-Id: I47d8536b9b2ed3dcc193d6e6b7f4573a4e22c9b3