Fix the gradient of `tf.gather()` when the indices are > 1-D.
The gradient function was previously generating an invalid
IndexedSlices, whereby `IndexedSlices.indices` tensor was not a
vector. This change reshapes the indices and gradient so that they can
correctly be interpreted as an IndexedSlices and applied to the
embedding variable.

Added a multi-dimensional gradient test in embedding_ops_test.py.

Fixes #505. Partially addresses #464.
Change: 110364370