tensorflow: relax the dimensionality check for tf.squeeze.
This CL relaxes the shape inference constraints on tf.squeeze.
Previously, trying to squeeze a dimension whose build-time shape
was unknown would result in the squeezed shape being completely
unknown. This causes errors on some otherwise working programs that
rely on TensorFlow's build-time shape inference.
With this change, it's assumed that, if there is an explicit squeeze_dims
argument to tf.squeeze() and the dims in question are not known at shape
inference time, the input dims will be assumed to be 1. If they are not,
an error will be thrown at runtime just as before.
Change: 113863596