Several changes to reduce the overhead and the cache footprint of dispatching nodes in the ExecutorState:
(1) For nodes with a small number of input and output edges, inlined
the input_type() and output_type() arrays into the NodeItem data
structure, which again helps with avoiding chasing random data in the Node*
data structure for most nodes.

(2) Cache the value of the virtual function call
params_.device()->RequiresRecordingAccessedTensors() in
ExecutorImpl::Initialize(), rather than calling it once for every node
execution on every step.

(3) Allocated a C++-style array directly for nodes_, and for
IterationState::input_tensors_ (rather than allocating a
std::vector<NodeItem> and a std::vector<Entry> to hold these,
respectively).  We know the size of the array at initialization time,
and this saves one extra level of indirection.

(4) Added a real constructor for IterationState, which avoided 3
replications of the same initialization code throughout the
executor.cc file.

(5) Augmented the NodeItem data structure with many additional fields
that can be precomputed about the Node or the Kernel for the node.
Many of these avoid an extra virtual function call or more per node
execution.  This helps cache locality considerably, since when we have
many distinct types of Kernels being executed in the graph, we pollute
the cache with many different virtual function tables.  (We still do
so somewhat because we invoke the kernel's Compute method, but we can perhaps
clean that up in a subsequent cl by saving a pointer to the Compute function
in the NodeItem as well).

(6) In OpKernelContext, cached the value of
params_.device()->RequiresRecordingAccessedTensors() by calling it
once on the constructor, rather than calling it once per
record_tensor_reference(...) call.  This also allows us to avoid a
lock acquisition when record_tensor_accesses_ is false when
OpKernelContext::retrieve_accessed_tensors is called, since we know
that we haven't recorded any accesses.

Together these improve the training performance of ptb_word_lm model
on my desktop using a Titan X GPU by about 1.2%.
Change: 114762889