Abstract the GCS/GFile code that event loading needs behind an io_wrapper module.
io_wrapper provides some functions that check whether the path is a GCS path and
calls the relevant functions from either gfile or gcs. This is *not* intended to
be a general-purpose interface; it only implements the things that are necessary
for loading events from GCS/GFile storage.

We're doing this because having the entire event loader stack care about the
difference between GCS and GFile is bad from an encapsulation perspective; this
way, we can present one consistent interface.
Change: 119427191