Add QOL improvements for tf.one_hot()
This commit adds a Python wrapper for tf.one_hot(), and introduce
several backwards compatible improvements, as mentioned in #1799:

1. The parameters `on_value` and `off_value` now have default values 1 and 0
respectively, as 32-bit integers. Because the most common use case of
one_hot() is to create Tensors for one-hot encoding, it will save users
some headache to not have to manually type in these values by hand each
time it is used. The parameters can still be given different values as
before, and this does not any previous behavior. In addition, the user
can choose to set only one of `on_value` or `off_value`, and the one
that is not set will maintain its default value.
2. `indices` and `depth` no longer have to be cast by hand as `int64` and
`int32` values, respectively. Previously, one_hot() would complain if
these values weren't of the exact right type- it now casts the values to
the appropriate type on its own.
3. The `one_hot()` function now has an additional input parameter,
dtype, which allows the user to specify the data type of the output
tensor

Added default values to one_hot function header

* `on_value` = 1
* `off_value` = 0
* `dtype` = `tf.float32`