fp16-enable all the componentwise ops.
This also includes updating parts of the Python test framework to handle fp16.
fp16 is too inaccurate to do numerical gradients unless a lot of care is taken,
so for fp16, we compare fp16 theoretical gradients to fp32 numerical ones.
This means that the gradient check doesn't also implicitly test the function
itself, so we will need to rely on the numpy tests for those.
Change: 119948035